+++ app/admin/layout.tsx
@@ -1,6 +1,6 @@
 import { cookies } from "next/headers";
 import Footer from "../footer";
-import { CONSTS } from "@/utils/constants";
+import { getRepositories } from "@/db/container";
 import { ADMIN_COOKIE_NAME, isAdminCookieValid } from "@/utils/auth";
 import { AdminHeader } from "./admin-header";
 
@@ -13,10 +13,11 @@ export default async function AdminLayout({
   const isAdmin = await isAdminCookieValid(
     cookieStore.get(ADMIN_COOKIE_NAME)?.value
   );
+  const { title } = await getRepositories().settings.get();
 
   return (
     <>
-      <AdminHeader title={CONSTS.TITLE} isAdmin={isAdmin} />
+      <AdminHeader title={title} isAdmin={isAdmin} />
       <main className="flex-1 flex flex-col px-3 lg:px-8 py-6 lg:pb-16">
         {children}
       </main>